Artist index
Shows a form to create a new artist, a list of artists from the database,
and a filter mechanism to quickly find the artist that you're looking for.
Artists can be created, updated, and deleted from this interface, and it will
also serve as the basis for the artist 'show' ('selected' in this app) when
there is more data to work with (albums, tracks, and favorites).